// src/app/model/data.ts
export interface Itodo {
    id: string;
    title: string;
}

export interface Istd {
    id: string;
    name: string;
    age: number;
    course: string;
    grade: string;
}

export interface Ipost {
    id: string;
    title: string;
    content: string;
}
export interface Imovie {
    id: string;
    title: string;
    description: string;
    rating: number;
    image: string;
}